Large enums are exported and generated by code to synchronize backend and frontend
