+++ internal_comptime/Cargo.toml
@@ -5,4 +5,5 @@ edition = "2024"
 
 [dependencies]
 phf = "0.12.1"
+serde = { version = "1.0.219", features = ["derive"] }
 shared_types = { path = "../shared_types" }
