+++ internal_comptime/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod cache;
 pub mod data;
 
+pub use cache::*;
 pub use data::*;
 pub use shared_types::*;
