+++ shared_types/Cargo.toml
@@ -5,5 +5,4 @@ edition = "2024"
 
 [dependencies]
 paste = "1.0.15"
-phf = { version = "0.12.1", features = ["macros"] }
 serde = { version = "1.0.219", features = ["derive"] }
